+++ src/candlewick/core/Components.cpp
@@ -1,0 +1,17 @@
+#include "Components.h"
+#include <algorithm>
+
+namespace candlewick {
+void updateTransparencyClassification(entt::registry &reg, entt::entity entity,
+                                      const MeshMaterialComponent &mmc) {
+  bool hasTransparency = std::ranges::any_of(mmc.materials, [](auto &material) {
+    return material.baseColor.w() < 1.0f;
+  });
+
+  if (hasTransparency) {
+    reg.remove<Opaque>(entity);
+  } else {
+    reg.emplace_or_replace<Opaque>(entity);
+  }
+}
+} // namespace candlewick